deeplogo_web/requirements.txt
absl-py==0.8.1
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
boto==2.49.0
boto3==1.10.28
botocore==1.13.28
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
cmake==3.15.3
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
Django==2.2.7
docutils==0.15.2
entrypoints==0.3
gast==0.3.2
gensim==3.8.1
grpcio==1.25.0
h5py==2.10.0
idna==2.8
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.2
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lab==5.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==3.0.5
more-itertools==8.0.2
MulticoreTSNE==0.1
mysqlclient==1.4.4
nbconvert==5.6.1
nbformat==5.0.3
notebook==6.0.2
numpy==1.17.4
olefile==0.46
opencv-python==4.1.2.30
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.0
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
s3transfer==0.2.1
scikit-learn==0.22.1
scipy==1.3.3
Send2Trash==1.5.0
simplejson==3.17.0
six==1.13.0
sklearn==0.0
smart-open==1.9.0
sqlparse==0.3.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
torch==1.3.1
torchvision==0.4.2
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
zipp==0.6.0
